Added Manual Testing.
feat: Added manual testing feature.

Added the ability to classify a line or many lines manually without the need for a dataset.

The only testing method that was available before this commit was the testing set method which requires taking the input from a CSV file. This can be troublesome and tiring if one wants to classify a single line or multiple lines (<10 for instance). This feature solves this problem all you have to do is to provide your input to the 'comment_text' field